services/fin_llm/requirements.lock.txt
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.12.1
PyYAML==6.0.3
Pygments==2.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.13.0
apache-tvm-ffi==0.1.9
astor==0.8.1
attrs==26.1.0
blake3==1.0.8
cachetools==7.0.5
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
compressed-tensors==0.12.2
cryptography==46.0.6
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-python==13.2.0
cupy-cuda12x==14.0.1
depyf==0.20.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.2
email-validator==2.3.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastapi==0.135.2
fastar==0.9.0
filelock==3.25.2
flashinfer-python==0.5.3
frozenlist==1.8.0
fsspec==2026.3.0
gguf==0.18.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx-sse==0.4.3
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ijson==3.5.0
interegular==0.3.3
jiter==0.13.0
jmespath==1.1.0
jsonschema-specifications==2025.9.1
jsonschema==4.26.0
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
mistral_common==1.10.0
model-hosting-container-standards==0.1.14
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
networkx==3.6.1
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.21.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl-libs-base==4.4.2
nvidia-cutlass-dsl==4.4.2
nvidia-ml-py==13.595.45
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-harmony==0.0.8
openai==2.30.0
opencv-python-headless==4.13.0.92
outlines_core==0.2.11
packaging==26.0
partial-json-parser==0.2.1.1.post7
pillow==12.1.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
py-cpuinfo==9.0.0
pybase64==1.4.3
pycountry==26.2.16
pycparser==3.0
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-json-logger==4.1.0
python-multipart==0.0.22
pyzmq==27.1.0
ray==2.54.1
referencing==0.37.0
regex==2026.3.32
requests==2.32.5
rich-toolkit==0.19.7
rich==14.3.3
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.17.1
sentencepiece==0.2.1
sentry-sdk==2.57.0
setproctitle==1.3.7
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.4
starlette==0.52.1
supervisor==4.3.0
sympy==1.14.0
tabulate==0.10.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.3
transformers==4.57.6
triton==3.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
vllm==0.13.0
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
xgrammar==0.1.27
yarl==1.23.0
